+++ Assets/LDtkUnity/CHANGELOG.md
@@ -2,6 +2,7 @@
 ##### April XX, 2022
 - Fixed entities not drawing their shape in the scene when their editor visual is set as `Tile`
 - Fixed the scales of some example entities to fit their area properly
+- Significantly updated documentation
 
 # 3.0.2
 ##### April 3, 2022